% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wind_Radii_Geom.R
\name{geocode_radii}
\alias{geocode_radii}
\title{Geocode Polygon Shapes}
\usage{
geocode_radii(point, arc = 0:360, radius = 1, conv = 1, scale = 1)
}
\arguments{
\item{point}{a numeric vector containing the longitude and latitude of the startin point of a polygon radius}

\item{arc}{a numeric vector containing the range of degrees over which to draw additional polygon points}

\item{radius}{a numeric object specifying the length of the radius of a polygon}

\item{conv}{a numeric object specifying a conversion factor, if the radius is not in meters}

\item{scale}{a numeric object between 0 and 1, specifying a scale by which to reduce the radius, if any}
}
\value{
a data frame of geocoded longitudes and latitude points, specifying the outline of a polygon
}
\description{
This function takes a point in space, given by a longitude and latitude, and projects out the
            outline of a radial polygon originating from the point. The outline it projects is made up of
            hundreds of additional points. The distance in projects it is based on the radius provided,
            and the extent of the polygon is determined by the degrees input into the variable "arc".
            If the radius is in any length of measurement other than meters, please also provide a
            convertion factor.
}
