% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wind_Radii_Geom.R
\docType{data}
\name{GeomHurricane}
\alias{GeomHurricane}
\title{Hurricane Geom}
\format{An object of class \code{GeomHurricane} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomHurricane
}
\description{
This is the Hurricane Wind Speed geom, which underlies the Hurricane Geom Layer function.
            Important parameters for this geom are x (longitude), y (latitude), r_ne, r_se, r_nw, r_sw
            (corresponding to wind distances in different directional quadrants), and scale_radii (which
            allows the user to scale down the wind effects, though this is optional).
}
\section{Aesthetics}{

\code{geom_hurricane} understands the following aesthetics (required in bold):
\itemize{
  \item \strong{x}
  \item \strong{y}
  \item \strong{r_ne}
  \item \strong{r_se}
  \item \strong{r_nw}
  \item \strong{r_sw}
  \item color
  \item fill
  \item size
  \item linetype
  \item alpha
  \item scale_radii
}
}

\keyword{datasets}
