% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wind_Radii_Extract_Tidy.R
\name{get_stormdata}
\alias{get_stormdata}
\title{Extract Data and Tidy}
\usage{
get_stormdata(file = "ebtrk_atlc_1988_2015.txt")
}
\arguments{
\item{file}{a character string specifying the name of a .txt file from which to read some data}
}
\value{
a data frame with a cleaned version of the input file data
}
\description{
This function simply extracts and tidies hurricane wind data from Colorado State University.
}
